+++ test/devServer.test.ts
@@ -57,13 +57,6 @@ suite('DevServerApi tests', () => {
 					key: 'test-flag',
 					value: true,
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [
-						{ id: '1', name: 'True', description: 'On', value: true },
-						{ id: '2', name: 'False', description: 'Off', value: false },
-					],
 				},
 			},
 			overrides: {
@@ -107,19 +100,11 @@ suite('DevServerApi tests', () => {
 					key: 'flag1',
 					value: true,
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 				'flag2': {
 					key: 'flag2',
 					value: 'test',
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			availableVariations: {},
@@ -153,10 +138,6 @@ suite('DevServerApi tests', () => {
 					key: 'target-flag',
 					value: 'found',
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			availableVariations: {},
@@ -285,13 +266,6 @@ suite('DevServerProvider tests', () => {
 					key: 'flag1',
 					value: true,
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [
-						{ id: '1', name: 'True', description: 'On', value: true },
-						{ id: '2', name: 'False', description: 'Off', value: false },
-					],
 				},
 			},
 			overrides: {
@@ -334,10 +308,6 @@ suite('DevServerProvider tests', () => {
 					key: 'overridden-flag',
 					value: 'original',
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			overrides: {
@@ -368,10 +338,6 @@ suite('DevServerProvider tests', () => {
 					key: 'normal-flag',
 					value: 'base-value',
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			overrides: {},
@@ -397,10 +363,6 @@ suite('DevServerProvider tests', () => {
 					key: 'flag1',
 					value: true,
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			overrides: {
@@ -431,10 +393,6 @@ suite('DevServerProvider tests', () => {
 					key: 'flag1',
 					value: true,
 					version: 1,
-					variation: 0,
-					trackEvents: false,
-					trackReason: false,
-					variations: [],
 				},
 			},
 			overrides: {},
@@ -456,9 +414,9 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
-				'flag2': { key: 'flag2', value: false, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
-				'flag3': { key: 'flag3', value: 'test', version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
+				'flag2': { key: 'flag2', value: false, version: 1 },
+				'flag3': { key: 'flag3', value: 'test', version: 1 },
 			},
 			overrides: {
 				'flag1': { value: false, version: 2 },
@@ -486,7 +444,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'test-flag': { key: 'test-flag', value: 'original', version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'test-flag': { key: 'test-flag', value: 'original', version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -524,7 +482,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -575,7 +533,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -596,8 +554,8 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
-				'flag2': { key: 'flag2', value: 'hello', version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
+				'flag2': { key: 'flag2', value: 'hello', version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -640,7 +598,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -666,7 +624,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -677,7 +635,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: false, version: 2, variation: 1, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: false, version: 2 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -704,7 +662,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
@@ -715,7 +673,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {
 				'flag1': { value: false, version: 2 },
@@ -744,7 +702,7 @@ suite('DevServerProvider tests', () => {
 			sourceEnvironmentKey: 'production',
 			context: {},
 			flagsState: {
-				'flag1': { key: 'flag1', value: true, version: 1, variation: 0, trackEvents: false, trackReason: false, variations: [] },
+				'flag1': { key: 'flag1', value: true, version: 1 },
 			},
 			overrides: {},
 			availableVariations: {},
